pyproject.toml
[tool.poetry]
name = "gen3"
version = "3.1.0"
description = "Gen3 Data Commons."
license = "Apache-2.0"
homepage = "https://gen3.org/"
repository = "https://github.com/uc-cdis/gen3"
authors = ["Fantix King <fantix@uchicago.edu>"]

[tool.poetry.dependencies]
python = "^3.7"
click = "^7.0"
fastapi = {version = "^0.42.0", optional = true}
uvicorn = {version = "^0.10.0", optional = true}
email-validator = {version = "^1.0", optional = true}
edgedb = {version = "^0.6.1", optional = true}
pypfb = {version = "^0.2.0", optional = true}
python-multipart = {version = "^0.0.5", optional = true}

[tool.poetry.dev-dependencies]

[tool.poetry.extras]
server = ["fastapi", "uvicorn", "email-validator", "edgedb", "pypfb", "python-multipart"]

[tool.poetry.scripts]
"gen3" = "gen3.cli:gen3"

[tool.poetry.plugins."gen3.cli"]
"server" = "gen3.server.cli [server]"

[tool.poetry.plugins."gen3.schema"]
"auth.user" = "gen3.auth.user:SCHEMA"

[tool.poetry.plugins."gen3.server"]
"auth.user" = "gen3.auth.user"
"submission.dict" = "gen3.submission.dictionary"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
